Fix non-determinism in tests
https://gitlab.com/gitlab-org/cluster-integration/gitlab-agent/-/jobs/794257153#L816

Because of concurrency an non-determinism there is a chance that GetObjectsToSynchronize() will be called another time even though the context has been cancelled and the application should finish. This cancellation is propagated via several goroutines and that is why there is no determinism. Test goroutine -> stager's Go() goroutine. Fix that by increasing the retry delay.